refactor: utils 폴더명 변경
기존 util에서 utils로 변경
